refactor: Context 분리
