+++ frontend/app/password-reset/page.tsx
@@ -1,23 +1,8 @@
 "use client";
 
-import { createContext, PropsWithChildren, useState } from "react";
+import { PropsWithChildren, useState } from "react";
 import PasswordResetFunnel from "@/components/password-reset/PasswordResetFunnel";
-
-export type PageStatus = "EMAIL_CODE" | "PASSWORD_SET" | "EMAIL_ENTRY";
-
-interface PageStatusContextProps {
-  pageStatus: PageStatus;
-  handlePageStatus: (status: PageStatus) => void;
-  verifiedEmail: string;
-  handleVerifiedEmail: (email: string) => void;
-}
-
-export const PageStatusContext = createContext<PageStatusContextProps>({
-  pageStatus: "EMAIL_CODE",
-  handlePageStatus: (status: PageStatus) => {},
-  verifiedEmail: "",
-  handleVerifiedEmail: (email: string) => {},
-});
+import { PageStatus, PageStatusContext } from "@/src/context";
 
 const PageStatusContextProvider = ({ children }: PropsWithChildren) => {
   const [pageStatus, setPageStatus] = useState<PageStatus>("EMAIL_ENTRY");
